refactor: split thumbnails.py file to multiple files
